Updated keypad.ino
Included new functions:
-turnOff();
-finalState();
-lcdPressed(char key);
-lcdDisp(int disp);
-currentState(char code);

included new variables:
laser
stateToPi

edited the following functions:
readKey(); 
void loop();

removed functions:
overrideState();

Summary:

Functionality has been configured and now it has a disarmed/arm state for the first line for lcd and inputted keys on 2nd line.
Failing to enter the password will decrement numAttempts, terminate if numAttempts is at 0 and change laser and state to 1.
Inputting the override code will increment numAttempts back to 5 and will disarm the turret, but will turn on the laser to easily differentiate that this is the override state. Entering a password wrong while in override state will reset the system to be armed and will keep the laser on.